requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.33.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
mangum==0.19.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
websockets==15.0.1
cryptography==44.0.2
pycryptodome==3.20.0
python-jose==3.3.0
